requirements.txt
Flask
python-dotenv
openai
langchain
langchain_openai